Add Token authentication

+++ MyCodeCamp/Controllers/SpeakersController.cs
@@ -22,19 +22,22 @@ namespace MyCodeCamp.Controllers
         private ILogger<SpeakersController> _logger;
         private IMapper _mapper;
         private ICampRepository _repo;
+        private UserManager<CampUser> _userManager;
 
         public SpeakersController(
             ICampRepository repo,
             ILogger<SpeakersController> logger,
-            IMapper mapper)
+            IMapper mapper,
+            UserManager<CampUser> userManager)
         {
             _repo = repo;
             _logger = logger;
             _mapper = mapper;
+            _userManager = userManager;
         }
 
         [HttpGet]
-        public IActionResult Get(string moniker, bool includeTalks = false)
+        public IActionResult Get(string moniker,bool includeTalks = false)
         {
             var speakers = includeTalks ? _repo.GetSpeakersByMonikerWithTalks(moniker) : _repo.GetSpeakersByMoniker(moniker);
 
@@ -42,10 +45,10 @@ namespace MyCodeCamp.Controllers
         }
 
         [HttpGet("{id}",Name = "SpeakerGet")]
-        public IActionResult Get(string moniker,int id, bool includeTalks = false)
+        public IActionResult Get(string moniker,int id,bool includeTalks = false)
         {
             var speaker = includeTalks ? _repo.GetSpeakerWithTalks(id) : _repo.GetSpeaker(id);
-            
+
 
             if (speaker == null) return NotFound();
             if (speaker.Camp.Moniker != moniker) return BadRequest("Speaker not in specified Camp");
@@ -55,6 +58,7 @@ namespace MyCodeCamp.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> Post(string moniker,[FromBody]SpeakerModel model)
         {
             try
@@ -65,11 +69,17 @@ namespace MyCodeCamp.Controllers
                 var speaker = _mapper.Map<Speaker>(model);
                 speaker.Camp = camp;
 
-                _repo.Add(speaker);
-                if (await _repo.SaveAllAsync())
+                var campUser = await _userManager.FindByNameAsync(this.User.Identity.Name);
+                if (campUser != null)
                 {
-                    var url = Url.Link("SpeakerGet",new { moniker = camp.Moniker,id = speaker.Id });
-                    return Created(url,_mapper.Map<SpeakerModel>(speaker));
+                    speaker.User = campUser;
+
+                    _repo.Add(speaker);
+                    if (await _repo.SaveAllAsync())
+                    {
+                        var url = Url.Link("SpeakerGet",new { moniker = camp.Moniker,id = speaker.Id });
+                        return Created(url,_mapper.Map<SpeakerModel>(speaker));
+                    }
                 }
             }
             catch (Exception ex)
@@ -82,6 +92,7 @@ namespace MyCodeCamp.Controllers
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> Put(string moniker,int id,[FromBody]SpeakerModel model)
         {
             try
@@ -89,6 +100,8 @@ namespace MyCodeCamp.Controllers
                 var speaker = _repo.GetSpeaker(id);
                 if (speaker == null) return NotFound();
                 if (speaker.Camp.Moniker != moniker) return BadRequest("Speaker and Camp do not match");
+
+                if (speaker.User.UserName != this.User.Identity.Name) return Forbid();
 
                 _mapper.Map(model,speaker);
 
@@ -106,13 +119,16 @@ namespace MyCodeCamp.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(string moniker, int id)
+        [Authorize]
+        public async Task<IActionResult> Delete(string moniker,int id)
         {
             try
             {
                 var speaker = _repo.GetSpeaker(id);
                 if (speaker == null) return NotFound();
                 if (speaker.Camp.Moniker != moniker) return BadRequest("Speaker and Camp do not match");
+
+                if (speaker.User.UserName != this.User.Identity.Name) return Forbid();
 
                 _repo.Delete(speaker);
